+++ SurgSim/Blocks/MassSpring2DRepresentation.h
@@ -54,8 +54,8 @@ public:
 	/// \note          |   |
 	/// \note   [0][0] *---* [1][0]
 	void init2D(const std::array<std::array<SurgSim::Math::Vector3d, 2>, 2> extremities,
-		unsigned int numNodesPerDim[2],
-		std::vector<unsigned int> nodeBoundaryConditions,
+		size_t numNodesPerDim[2],
+		std::vector<size_t> nodeBoundaryConditions,
 		double totalMass,
 		double stiffnessStretching, double dampingStretching,
 		double stiffnessBending, double dampingBending,
@@ -67,19 +67,19 @@ private:
 	/// \param numNodesPerDim The number of nodes on the 2 dimensions
 	/// \param stiffness, damping The spring parameters
 	void init2DStretchingSprings(const std::shared_ptr<SurgSim::Math::OdeState> state,
-		unsigned int numNodesPerDim[2], double stiffness, double damping);
+		size_t numNodesPerDim[2], double stiffness, double damping);
 	/// Helper method to initialize/add all bending springs on a 2D structure
 	/// \param state The state to initialize the springs with (rest lengths calculation)
 	/// \param numNodesPerDim The number of nodes on the 2 dimensions
 	/// \param stiffness, damping The spring parameters
 	void init2DBendingSprings(const std::shared_ptr<SurgSim::Math::OdeState> state,
-		unsigned int numNodesPerDim[2], double stiffness, double damping);
+		size_t numNodesPerDim[2], double stiffness, double damping);
 	/// Helper method to initialize/add all face diagonal springs on a 2D structure
 	/// \param state The state to initialize the springs with (rest lengths calculation)
 	/// \param numNodesPerDim The number of nodes on the 2 dimensions
 	/// \param stiffness, damping The spring parameters
 	void init2DFaceDiagonalSprings(const std::shared_ptr<SurgSim::Math::OdeState> state,
-		unsigned int numNodesPerDim[2], double stiffness, double damping);
+		size_t numNodesPerDim[2], double stiffness, double damping);
 };
 
 };  // namespace Blocks
